Refactor code and address reviewer feedback
