+++ example/ck_tile/35_batched_transpose/README.md
@@ -21,7 +21,7 @@ make tile_example_batched_transpose_2d -j
 make tile_example_batched_transpose_4d -j
 ```
 
-This will generate the executable:
+This will generate the executables:
 
 ```
 build/bin/tile_example_batched_transpose_2d
@@ -38,8 +38,14 @@ build/bin/tile_example_batched_transpose_4d
 
 #### Arguments:
 
-- `-H` : Input height size (default: 16)
-- `-W` : Input width size (default: 16)
+```
+args:
+          -v    whether do CPU validation or not (default:1)
+         -pr    input data type. fp16/fp32 (representing 8/16/32 bit data) (default:fp16)
+       -seed    seed to be used, -1 means random every time (default:-1)
+          -H    input height size. (default:16)
+          -W    input width size.  (default:16)
+```
 
 ### 4D Transpose (NCHW -> NHWC)
 
@@ -49,11 +55,15 @@ build/bin/tile_example_batched_transpose_4d
 
 #### Arguments:
 
-- `-N`  : Input batch size (default: 2)
-- `-C`  : Input channel size (default: 16)
-- `-H`  : Input height size (default: 1)
-- `-W`  : Input width size (default: 16)
-- `-layout_in`  : Input tensor layout (default: NCHW)
-- `-layout_out` : Output tensor layout (default: NHWC)
-- `-v`  : Enable CPU validation (default: 1)
-- `-seed` : Random seed (-1 means random every time, default: -1)
+```
+args:
+          -v    whether do CPU validation or not (default:1)
+         -pr    input data type. fp16/fp32 (representing 8/16/32 bit data) (default:fp16)
+       -seed    seed to be used, -1 means random every time (default:-1)
+          -N    input batch size.  (default:2)
+          -C    input channel size. (default:16)
+          -H    input height size. (default:1)
+          -W    input width size.  (default:16)
+  -layout_in    input tensor data layout - NCHW by default (default:NCHW)
+ -layout_out    output tensor data layout - NHWC by default  (default:NHWC)
+```
