+++ example/ck_tile/35_batched_transpose/batched_transpose.hpp
@@ -10,16 +10,15 @@
 
 #pragma once
 
-struct batched_transpose_4d_trait
+struct batched_transpose_trait
 {
     std::string type;
-    std::string layout;
 };
 
-struct batched_transpose_4d_kargs : public ck_tile::BatchedTransposeHostArgs
+struct batched_transpose_kargs : public ck_tile::BatchedTransposeHostArgs
 {
 };
 
-float batched_transpose_4d(batched_transpose_4d_trait t,
-                           batched_transpose_4d_kargs a,
-                           ck_tile::stream_config s);
+float batched_transpose(batched_transpose_trait t,
+                        batched_transpose_kargs a,
+                        ck_tile::stream_config s);
